Residual operation improve
Bit operation only needs an instruction cycle to complete, and most of the C compiler "%" operation is to call subroutines to complete, the code is long, slow execution speed. In general, as long as the remainder of 2n square is sought, the method of bit operation can be used instead.